Cleaned up __main__.py and bumped version, moved old script to v0.1.py
